requirements.txt
cffi==1.14.3
click==7.1.2
cryptography==3.1
Flask==1.1.2
flask-paginate==0.7.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.2
Pillow==7.2.0
pycparser==2.20
qrcode==6.1
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3